requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.7.0
google-generativeai==0.3.1
psycopg[binary]==3.1.14
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
sqlalchemy==2.0.23
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4